moved resource methods into entity
